hw7 Final

+++ Binomial.java
@@ -13,7 +13,7 @@ public class Binomial {
 	}
 
 	// Computes the Binomial function, basic version.
-	public static int binomial1(int n, int k) {
+	public static long binomial1(int n, int k) {
 		if (k > n) {
 			return 0;
 		}
@@ -24,8 +24,8 @@ public class Binomial {
 	}
 
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
-		int[][] memo = new int[n + 1][k + 1];
+	public static long binomial(int n, int k) {
+		long[][] memo = new long[n + 1][k + 1];
 		for (int i = 0; i < memo.length; i++) {
 			for (int j = 0; j < memo[0].length; j++) {
 				memo[i][j] = -1;
@@ -34,7 +34,7 @@ public class Binomial {
 		return binomial(n, k, memo);
 	}
 
-	private static int binomial(int n, int k, int[][] memo) {
+	private static long binomial(int n, int k, long[][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
